requirements.txt
# Deep Learning & Computer Vision
torch>=2.6.0
torchvision>=0.21.0
opencv-python>=4.8.0
scikit-image>=0.23.0
Pillow>=10.0.0

# Transformers & Models
transformers>=4.35.0
timm>=0.9.12

# Vector Search & Indexing
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support
numpy>=1.24.0

# Hashing & Similarity
imagehash>=4.3.1
scikit-learn>=1.3.0

# Database & Storage
h5py>=3.10.0
lmdb>=1.4.1

# GUI
PyQt6>=6.6.0
pyqtgraph>=0.13.3
matplotlib>=3.8.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
joblib>=1.3.0
psutil>=5.9.0

# Development
pytest>=7.4.0
black>=23.11.0
flake8>=6.1.0  